v2/smart_test_llm/requirements.txt
openai==1.17.1
python-dotenv
langchain==0.1.16
langsmith==0.1.47
langchain-openai==0.1.3
langchain-chroma==0.1.0
langchain-community
adbutils==2.3.1
xmltodict==0.13.0
beautifulsoup4==4.12.3
lxml==5.2.1
uiautomator2==3.0.8
# jsonpickle==2.0.0
pillow
# sentence-transformers
tenacity
matplotlib
pypdf
esprima
tree_sitter
tree_sitter_languages
unstructured
